// src/products/dto/product.dto.ts
export class CDNResponseDto {
    items:CDNItem[]

}

export class CDNItem {
    fields:{
        sku: string;
        name: string;
        brand: string;
        model: string;
        category: string;
        color: string;
        price: number;
        currency: string;
        stock: number;
    }
} 